yubi update

+++ yubikey.zsh
@@ -4,12 +4,12 @@ alias gpg=gpg2
 #compdef gpg2=gpg
 
 # Yubikey , GPG, SSH
-if [ -x /usr/bin/gpg-agent ]; then
-#  echo "gpg-agent exists"
+if [ -x /usr/bin/gpg-agent ] && [ -x /usr/lib/gnupg2/scdaemon ]; then
+  #echo "gpg-agent exists"
   if [[ ! $(ps x | grep gpg-agent | grep ssh | grep -v grep) ]]; then
-#    echo "no gpg-agent exists"
+    #echo "no gpg-agent exists"
     while [[ $(ps x | grep gpg-agent | grep -v grep) ]]; do
-#      echo "trying to kill the agent"
+      #echo "trying to kill the agent"
       killall gpg-agent > /dev/null 2>&1
     done
     #if [ ! -f $HOME/.gpg-agent-info ]; then
